requirements.txt
anthropic>=0.34.0
fastapi>=0.115.0
uvicorn>=0.30.6
click>=8.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
portalocker>=2.8.0
rich>=13.7.0
